Added c to b
